Fixed misalignment, added throttle position
Changed interpolation method so it interpolates the data according to a whole number frequency; large batches of data shouldn't get misaligned like they would have. Interpolation and data processing methods return the HZ of a processed channel. I added a throttle data file and a processing method for that. I reorganized the code to be better to use.