Full support for our CANBus + brake transformation
Added full support for EVERY channel in our CAN Bus files. I also added the processing for our brake data. I think we should remove most of the channels that were added that weren't in Reid's original list if we can because they seem sort of useless. WARNING: THIS CODE IS UGLY!!